Move all lowering patterns to one file
There is no reason to have several tiny files. If we ever hit lengths
that are problematic, typically >10k LoC, we can split.
